% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-auc.r
\name{calc_auc}
\alias{calc_auc}
\title{Area under a curve (AUC)}
\usage{
calc_auc(x, y)
}
\arguments{
\item{x}{the variable plotted on the x-axis of the curve plot, e.g. for a plot with an ROC curve, x-axis is the false positive rate}

\item{y}{the varialbe plotted on the y-axis of the curve plot, e.g. for a plot with an ROC curve, y-axis is the true positive rate}
}
\description{
This function calculates the area under an ROC curve or the area under a precision-recall curve.
}
\examples{
library(tidyverse)
library(broom)
library(tidyroc)

# get `biopsy` dataset from `MASS`
data(biopsy, package = "MASS")

# change column names from `V1`, `V2`, etc. to informative variable names
colnames(biopsy) <-
  c(
    "ID",
    "clump_thickness",
    "uniform_cell_size",
    "uniform_cell_shape",
    "marg_adhesion",
    "epithelial_cell_size",
    "bare_nuclei",
    "bland_chromatin",
    "normal_nucleoli",
    "mitoses",
    "outcome"
  )

# fit a logistic regression model to predict tumour type
glm(outcome ~ clump_thickness + uniform_cell_shape,
  family = binomial,
  data = biopsy
) \%>\%
  augment() \%>\% # use broom to add glm output to the #' original data frame
  make_roc(predictor = .fitted, known_class = outcome) \%>\% # get values to plot an ROC curve
  summarise(auc = calc_auc(x = fpr, y = tpr)) # calculate the area under an ROC curve
}
