% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure-perf.r
\name{calc_tpr}
\alias{calc_tpr}
\title{Measure model performance}
\usage{
calc_tpr(pred_values, pos_pred, pos)
}
\arguments{
\item{data}{data-frame that contains fitted values and known outcomes}

\item{predictor}{column in `data` that contains fitted values}

\item{known_class}{column in `data` that contains true or actual classification}
}
\description{
This function evaluates a binary classification model. The function calculates true positive rate, false positive rate, true negative rate, false negative rate, and positive predictive value. These values are added to the data-frame provided in `data`.
}
\examples{
library(tidyverse)
library(broom)
library(tidyroc)

# get `biopsy` dataset from `MASS`
data(biopsy, package = "MASS")

# change column names from `V1`, `V2`, etc. to informative variable names
colnames(biopsy) <-
 c("ID",
   "clump_thickness",
   "uniform_cell_size",
   "uniform_cell_shape",
   "marg_adhesion",
   "epithelial_cell_size",
   "bare_nuclei",
   "bland_chromatin",
   "normal_nucleoli",
   "mitoses",
   "outcome")

# fit a logistic regression model to predict tumour type
glm(outcome ~ clump_thickness + uniform_cell_shape,
   family = binomial,
   data = biopsy
) \%>\%
 augment() \%>\% # use broom to add glm output to the original data frame
 measure_perf(predictor = .fitted, known_class = outcome)

}
