% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-conf-matrix.r
\name{make_conf_matrix}
\alias{make_conf_matrix}
\title{Receiver operating characteristic (ROC) curve}
\usage{
make_conf_matrix(data, ...)
}
\arguments{
\item{probabilities}{}
}
\description{
this function calculates true positive rate and false positive rate to make an ROC curve
}
\examples{


}
