% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-conf-matrix.r
\name{make_conf_matrix_ungrouped}
\alias{make_conf_matrix_ungrouped}
\title{Receiver operating characteristic (ROC) curve}
\usage{
make_conf_matrix_ungrouped(data, key, cutoff, predictor, known_class)
}
\arguments{
\item{probabilities}{}
}
\description{
this function calculates true positive rate and false positive rate to make an ROC curve
}
\examples{

}
